requirements.txt
pandas
seaborn
djangorestframework